README.md
Creating mandala pattern using p5.js.
This is challenging for me to create a pattern using this kind of application. I created a pattern design that reflects our way of life. It'll starts with a small circle that represents a new beginnings in life. Next is the lines connected to the circle is the    obstacles we need to face in order to go on the next journey. The small circles that surrounds is inspirations that no matter what challenges and obstacles that happened, they are always there to motivates you. The thorns allign with big circles and small circles inside means the positive and negative thoughts that comes to our mind especially when we have problems that need to accomplish. Lastly the big circles, each of them group with five to form flowers. So, I choose this design because flowers symbolizes adoration, dedication, spiritual attainment, flexibility, and opportunity. Also symbolic of good luck, wealth, happiness, strength, and ambition.
 That no matter what happens in life there is hope on us. So I will never give up in order to achieve my goals, learn from my mistakes and be positive in life.